+++ src/main/kotlin/Permissions.kt
@@ -2,7 +2,7 @@ package org.kamiblue.botkt
 
 import org.kamiblue.botkt.helpers.StringHelper.toHumanReadable
 import net.ayataka.kordis.entity.message.Message
-import org.kamiblue.botkt.Send.error
+import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.Colors
 
 object Permissions {
