+++ src/main/kotlin/commands/RoleInfoCommand.kt
@@ -3,7 +3,7 @@ package org.kamiblue.botkt.commands
 import net.ayataka.kordis.entity.message.Message
 import net.ayataka.kordis.entity.server.role.Role
 import org.kamiblue.botkt.*
-import org.kamiblue.botkt.Send.error
+import org.kamiblue.botkt.utils.MessageSendUtils.error
 import org.kamiblue.botkt.utils.Colors
 
 object RoleInfoCommand : Command("roleinfo") {
